pyproject.toml
[project]
name = "ignition"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "albumentations>=1.3.0",
    "flatten-dict>=0.4.2",
    "hydra-core>=1.3.2",
    "image-dataset-viz>=0.2.2",
    "isort>=6.0.1",
    "jupyter>=1.1.1",
    "monai[einops,gdown,h5py,huggingface-hub,imagecodecs,itk,matplotlib,mlflow,nibabel,ninja,optuna,pandas,pillow,pydicom,tifffile,tqdm,transformers]>=1.5.0",
    "pytest>=8.4.1",
    "pytorch-ignite>=0.4.8",
    "tensorboard>=2.20.0",
    "torch>=1.10.2",
    "torchvision>=0.11.3",
]

[tool.uv]
package = true
